MSVC compilation fix
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@4812 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ creature.cpp
@@ -299,7 +299,7 @@ void Creature::onWalk(Direction& dir)
 	if(r > drunk)
 		return;
 
-	int32_t tmp = std::floor(drunk / 5);
+	int32_t tmp = std::floor((double)drunk / 5.);
 	if(r <= tmp)
 		dir = NORTH;
 	else if(r <= tmp * 2)
